Blocks: Merge paragraph and cover text blocks
