--- blocks/library/cover-text/editor.scss
@@ -1,3 +1,0 @@
-.wp-block-cover-text p.blocks-editable__tinymce {
-	font-size: inherit;
-}
