--- blocks/library/cover-text/style.scss
@@ -1,12 +1,0 @@
-.wp-block-cover-text {
-
-	&.has-background {
-		padding: 20px 30px;
-	}
-
-	p {
-		font-size: inherit !important;
-		max-width: $visual-editor-max-width - ( $block-mover-padding-visible );
-		margin: 0 auto;
-	}
-}
