+++ blocks/library/index.js
@@ -17,7 +17,6 @@ import './freeform';
 import './latest-posts';
 import './categories';
 import './cover-image';
-import './cover-text';
 import './shortcode';
 import './text-columns';
 import './verse';
