+++ blocks/library/paragraph/style.scss
@@ -14,3 +14,7 @@ p.has-drop-cap {
 p.has-drop-cap:not( :focus )  {
 	overflow: hidden;
 }
+
+p.has-background {
+	padding: 20px 30px;
+}
