--- blocks/test/fixtures/core__cover-text.json
@@ -1,17 +1,0 @@
-[
-    {
-        "uid": "_uid_0",
-        "name": "core/cover-text",
-        "isValid": false,
-        "attributes": {
-            "content": [
-                [
-                    "Text with color."
-                ]
-            ],
-            "dropCap": false,
-            "backgroundColor": "#fcb900"
-        },
-        "originalContent": "<div style=\"background-color:#fcb900\"><p>Text with color.</p></div>"
-    }
-]
