--- blocks/test/fixtures/core__cover-text.serialized.html
@@ -1,5 +1,0 @@
-<!-- wp:core/cover-text {"backgroundColor":"#fcb900"} -->
-<div style="background-color:#fcb900">
-    <p>Text with color.</p>
-</div>
-<!-- /wp:core/cover-text -->
